移动 feign 代码 到 授权服务
对于基于反应式 webflux 的程序，feign不在支持，不再考虑使用
feign-reactive
仅仅保留 授权服务 的 feign模块
其他服务将使用 webclient 重写

+++ oauth2-auth-server/src/main/kotlin/com/qingzhu/oauth2authserver/security/StaffUserDetailsService.kt
@@ -23,7 +23,7 @@ class StaffUserDetailsService : UserDetailsService {
     override fun loadUserByUsername(username: String?): User {
         val request = (RequestContextHolder.getRequestAttributes() as ServletRequestAttributes).request
         val org_id = request.getParameter("org_id")
-        val innerUser = staffService.findFirstByUsername(org_id.toLong(), username)
+        val innerUser = staffService.findFirstByUsername(org_id.toInt(), username)
         return innerUser?.toMyUser() ?: throw UsernameNotFoundException("$username not found")
     }
